requirements.txt
requests==2.31.0
pycryptodomex==3.20.0